Move the shared_ptr atomic helpers into atomic_compat.h
